+++ lib/mongoid/persistable/renamable.rb
@@ -4,7 +4,7 @@ module Mongoid
 
     # Defines behaviour for $rename operations.
     #
-    # @since 2.0.0
+    # @since 4.0.0
     module Renamable
       extend ActiveSupport::Concern