+++ lib/mongoid/persistable/unsettable.rb
@@ -4,7 +4,7 @@ module Mongoid
 
     # Defines behaviour for $unset operations.
     #
-    # @since 2.0.0
+    # @since 4.0.0
     module Unsettable
       extend ActiveSupport::Concern